Fixes compiler warnings on Mac
Former-commit-id: 29c0e8bee500451a5764bd91aa6f9ff7e3d03599

+++ cocos2d-ui/CCTableView.m
@@ -294,7 +294,7 @@
     if (!_dataSource) return;
     
     // Resize the content node
-    int numRows = [_dataSource tableViewNumberOfRows:self];
+    NSUInteger numRows = [_dataSource tableViewNumberOfRows:self];
     float layerHeight = 0;
     
     if ([_dataSource respondsToSelector:@selector(tableView:heightForRowAtIndex:)])
